refactor out key generation factory
